+++ src/main/kotlin/net/andrecarbajal/telegramdiscountsbot/scrapping/Scrapping.kt
@@ -19,11 +19,12 @@ private fun scrapeWebsite(
     nameCssSelector: String,
     priceCssSelector: String,
     offerCssSelector: String
-): Triple<String, String, String?>? {
+): List<String?>? {
+    val waitCssSelector = "div.product-detail-content"
+    val imageCssSelector = "img.ngxImageZoomThumbnail"
+
     val options = FirefoxOptions()
-    // Execute the browser without UI
     options.addArguments("--headless")
-    // Execute the browser without security
     options.addArguments("--no-sandbox")
 
     val driver: WebDriver = try {
@@ -35,15 +36,16 @@ private fun scrapeWebsite(
 
     return try {
         driver.get(url)
-        val wait = WebDriverWait(driver, Duration.ofMinutes(10))
-        wait.until(ExpectedConditions.presenceOfElementLocated(By.cssSelector("div.product-detail-content")))
+        val wait = WebDriverWait(driver, Duration.ofMinutes(5))
+        wait.until(ExpectedConditions.presenceOfElementLocated(By.cssSelector(waitCssSelector)))
 
         val doc: Document = Jsoup.parse(driver.pageSource!!)
         val name = doc.select(nameCssSelector).firstOrNull()?.text() ?: "Name not found"
         val price = doc.select(priceCssSelector).firstOrNull()?.text() ?: "Price not found"
         val offer = doc.select(offerCssSelector).getOrNull(1)?.text()
+        var img = doc.select(imageCssSelector).firstOrNull()?.attr("src")
 
-        Triple(name, price, offer)
+        listOf(name, price, offer, img)
     } catch (e: Exception) {
         logger.error("Error occurred during scraping", e)
         null
@@ -59,7 +61,7 @@ fun scrappingMifarma(url: String): List<String?>? {
         "div.col-xs-4.col-sm-2.col-md-6.col-lg-4.text-right.d-flex.align-items-center.justify-content-end.price-amount",
         "div.col-xs-4.col-sm-2.col-md-6.col-lg-4.text-right.d-flex.align-items-center.justify-content-end.price-amount"
     ) ?: return null
-    return listOf("MiFarma", result.first, result.second, result.third)
+    return listOf("MiFarma", result[0], result[1], result[2], result[3])
 }
 
 fun scrappingInkaFarma(url: String): List<String?>? {
@@ -69,5 +71,5 @@ fun scrappingInkaFarma(url: String): List<String?>? {
         "div[class*='col-lg-4']",
         "div.col-xs-5.col-sm-2.col-md-6.col-lg-4.text-right.d-flex.align-items-center.justify-content-end.price-amount"
     ) ?: return null
-    return listOf("InkaFarma", result.first, result.second, result.third)
+    return listOf("InkaFarma", result[0], result[1], result[2], result[3])
 }