+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
 # Release Notes
 
-- Adding seconds to nextRun
+- If the image of the product is loaded on scrapping, the image will be sent with the information of the offer.
+- The URL of the product will be sent with the information of the offer.